extract/requirements.txt
dicttoxml
PyYAML
guess_language-spirit